pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anonymcp"
version = "0.1.0"
description = "An MCP server for data governance — PII detection, anonymization, classification, and audit powered by Microsoft Presidio."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Frank" },
]
keywords = ["mcp", "pii", "presidio", "data-governance", "anonymization", "privacy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "mcp>=1.2.0",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "spacy>=3.7.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "aiofiles>=24.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
transformers = ["presidio-analyzer[transformers]"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
    "mypy>=1.10",
    "pre-commit>=3.0",
]
all = ["anonymcp[transformers,dev]"]

[project.scripts]
anonymcp = "anonymcp.server:main"

[project.urls]
Homepage = "https://github.com/frankkyazze/anonymcp"
Repository = "https://github.com/frankkyazze/anonymcp"
Issues = "https://github.com/frankkyazze/anonymcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/anonymcp"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
